+++ srcs/init_utils/parse_cubfile/parse_map_utils/check_last_map.c
@@ -6,13 +6,13 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 19:20:57 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/12/17 19:21:06 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/12/17 21:46:07 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static bool	is_map_element(const char *str)
+static bool	is_map_line(const char *str)
 {
 	const char	*identifier = " 10NWES";
 	size_t		i;
@@ -37,32 +37,39 @@ static bool	is_map_element(const char *str)
 	return (true);
 }
 
+static int	check_continue_of_map(bool is_map_line, const char *line)
+{
+	if (is_map_line == true && *line != '\n')
+	{
+		ft_printf("%s%s\nNG line: [", ERR_PROMPT, EMSG_MAP_CHAR);
+		print_until_ch(STDOUT_FILENO, line, '\n');
+		ft_printf("]\n");
+		return (EXIT_FAILURE);
+	}
+	if (is_map_line == false && *line != '\n')
+	{
+		ft_printf("%s%s\n", ERR_PROMPT, EMSG_MAP_NOT_LAST);
+		return (EXIT_FAILURE);
+	}
+	return (EXIT_SUCCESS);
+}
+
 int	check_last_map(const char *line, t_parse *parse)
 {
-	bool	is_checking_in_map_element;
+	bool	is_find_map;
 
 	while (line != NULL)
 	{
-		is_checking_in_map_element = false;
-		while (is_map_element(line) == true)
+		is_find_map = false;
+		while (is_map_line(line) == true)
 		{
-			is_checking_in_map_element = true;
+			is_find_map = true;
 			line = find_next_line(line);
 			if (line == NULL)
 				return (EXIT_SUCCESS);
 		}
-		if (is_checking_in_map_element == true && *line != '\n')
-		{
-			ft_printf("%s%s\nNG line: [", ERR_PROMPT, EMSG_MAP_CHAR);
-			print_until_ch(STDOUT_FILENO, line, '\n');
-			ft_printf("]\n");
+		if (check_continue_of_map(is_find_map, line) != EXIT_SUCCESS)
 			return (EXIT_FAILURE);
-		}
-		if (is_checking_in_map_element == false && *line != '\n')
-		{
-			ft_printf("%s%s\n", ERR_PROMPT, EMSG_MAP_NOT_LAST);
-			return (EXIT_FAILURE);
-		}
 		line = find_next_line(line);
 	}
 	(void)parse;
